Rename admin.hierarchy to admin.site
