+++ misago/admin/__init__.py
@@ -1,4 +1,4 @@
-from .hierarchy import site
+from .site import site
 from .urlpatterns import urlpatterns
 from .discoverer import discover_misago_admin
